starting with the moloch v2 contructor
